refactor(web): share CodeMirror highlighting primitives
