+++ web/src/lib/highlighting/__tests__/code-highlight-types.logic.test.ts
@@ -1,6 +1,22 @@
 import { describe, expect, it } from 'vitest';
 
-import { plainCodeSegments } from '../code-highlight-types';
+import { appendCodeHighlightSegment, plainCodeSegments } from '../code-highlight-types';
+
+describe('appendCodeHighlightSegment', () => {
+	it('merges adjacent segments with the same normalized class', () => {
+		const segments = [{ text: 'const', className: 'cm-code-keyword' }];
+
+		appendCodeHighlightSegment(segments, ' value', 'cm-code-keyword');
+		appendCodeHighlightSegment(segments, ' = 1', '');
+		appendCodeHighlightSegment(segments, ';', null);
+		appendCodeHighlightSegment(segments, '', 'cm-code-title');
+
+		expect(segments).toEqual([
+			{ text: 'const value', className: 'cm-code-keyword' },
+			{ text: ' = 1;', className: null },
+		]);
+	});
+});
 
 describe('plainCodeSegments', () => {
 	it('returns no segments for empty text', () => {
