+++ web/src/lib/highlighting/code-highlight-types.ts
@@ -3,6 +3,23 @@ export interface CodeHighlightSegment {
 	className: string | null;
 }
 
+export function appendCodeHighlightSegment(
+	segments: CodeHighlightSegment[],
+	text: string,
+	className: string | null,
+): void {
+	if (!text) return;
+
+	const normalizedClassName = className || null;
+	const previous = segments.at(-1);
+	if (previous?.className === normalizedClassName) {
+		previous.text += text;
+		return;
+	}
+
+	segments.push({ text, className: normalizedClassName });
+}
+
 export function plainCodeSegments(text: string): CodeHighlightSegment[] {
 	return text ? [{ text, className: null }] : [];
 }
